requirements.txt
playwright==1.41.0
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-api-python-client==2.116.0
pyyaml==6.0.1
python-dotenv==1.0.1
schedule==1.2.1
